Faster prime testing
Prime testing now uses the Miller-Rabin test instead of brute force. The
code for the Miller-Rabin test was originally written for a project in
the first term of my Masters degree. The primality test is now random,
so there is a chance that it will make a mistake, though Wikipedia says
that with the 10 random trials, the odds of it failing are less than one
in a million. Also, much larger numbers can be tested now! Before,
Python would run out of memory with large numbers trying to create
range(2, int(math.sqrt(n)) + 1). Oops.